+++ examples/document-client.ts
