+++ package.json
@@ -12,7 +12,7 @@
     "test": "TZ=UTC jest",
     "prepare": "husky install"
   },
-  "main": "src/lib",
+  "main": "src/runtime",
   "engines": {
     "node": "16.x",
     "npm": "8.x"
@@ -39,7 +39,6 @@
     "@types/node": "^16.18.2",
     "@typescript-eslint/eslint-plugin": "^5.41.0",
     "@typescript-eslint/parser": "^5.41.0",
-    "babel-plugin-module-resolver": "^4.1.0",
     "eslint": "^8.26.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-babel": "^5.3.1",
