+++ src/codegen/actions/plugin.ts
@@ -83,15 +83,14 @@ ${tableTypes
     assert(info?.outputFile, 'info.outputFile is required');
 
     const isExample = !!process.env.IS_EXAMPLE;
+    const runtimeModuleId = isExample ? '../../..' : '@ianwremmel/data';
 
     return {
       content,
       prepend: [
         `import {ConditionalCheckFailedException, ConsumedCapacity, ItemCollectionMetrics} from '@aws-sdk/client-dynamodb';`,
         `import {DeleteCommand, GetCommand, QueryCommand, UpdateCommand} from '@aws-sdk/lib-dynamodb'`,
-        isExample
-          ? `import {assert, DataIntegrityError, NotFoundError, OptimisticLockingError} from '../../..'`
-          : `import {assert, DataIntegrityError, NotFoundError, OptimisticLockingError} from '@ianwremmel/data'`,
+        `import {assert, DataIntegrityError, NotFoundError, OptimisticLockingError} from '${runtimeModuleId}'`,
         `import {v4 as uuidv4} from 'uuid'`,
         `import {ddbDocClient} from "${path.relative(
           path.resolve(process.cwd(), path.dirname(info.outputFile)),
